+++ Core/Core/Serialisation/SpeckleSerializerException.cs
